fix: Fixed the issue on completing the offer service test script.
